Use CITEXT
Whenever you hit a problem while working with PostgreSQL, search
for extensions. There are plenty of different extensions out there
and you can probably find one that helps you.

In this case, we were able to solve case insensitive texts by simply
replacing TEXT to CITEXT.

But, that's not enough for color/image processing. We need a better
way of storing the data.