Introduce the Extended color node
In postgresql source code, "node" is a very common base structure.
It has some basic properties such as having copy/equal/read/out
functions.

Basically, postgresql's query/expression trees are made up of "node"s.
Some examples are "Plan", "Limit" or "Const". To see all, read the code
in include/nodes/nodes.h

PostgreSQL also allows extending the "nodes" in the sense that having
the ability to add new node types (a.k.a., ExtensibleNode).

With this commit, we're adding a new type of node, called
PgColorExtendedNode. Our primary motivation is to use this node to
pass information from the planner to the executor.